requirements.txt
openai
faiss-cpu
numpy
pypdf
python-dotenv